+++ src/game/tests.rs
@@ -971,6 +971,7 @@ mod lost_caverns_of_ixalan;
 mod marvel_super_heroes;
 mod playtest_regressions;
 mod plot_exile;
+mod ral_crackling_wit;
 mod ramos_dragon_engine;
 mod secrets_of_strixhaven;
 mod spider_man;